Moved registry to a separate file
